huge speedup with cost calculations for large n, m. finally realised why the matrix mult was so slow! it was multiplying a logical, not a double, matrix.

+++ Population_based_heuristics/calc_machine_costs.m
@@ -6,7 +6,7 @@ function [machine_cost_mat] = calc_machine_costs(jobs_array_aug, pop_mat, ...
     jobs_array_transpose = jobs_array_aug';
     machine_cost_mat = zeros(size(pop_mat,1), num_machines);
     for j = 1:num_machines
-        indicator_mat = (pop_mat == j);
+        indicator_mat = double((pop_mat == j));
         machine_cost_mat(:,j) = indicator_mat * jobs_array_transpose;
     end
 end
